How do I migrate the database?
Analogy: migrating a database is like convincing a heard of cats to leave one house and move to another.  It takes a lot of tricky manuvering. so you hire another person (with a rake) to do it for you.

[like this:](https://gist.github.com/tararoys/b62f1e0a29c1ffee7a93/41951c33190963bc1a40e0d20cc21b91e71dae0a)

With the command

   rake db:migrate

Note: I assumed that you couldn't add test.sqlite3 or development.sqlite3 to my git because they were empty files.  Turns out you can, but git ignores them because they are in the gitignore file.  Fancy that!