How to I see what happened to the database when I dropped, created, and remigrated it?
Answer: the file development.log keeps track of all of the stuff you did.

The development.log is like the photo album your parents keep of you as you grow up.  It shows how the app changes over time.  It doesn't capture everyting, but it does give you a sense of what changed and where it changed.