How do I add a route to a rails app?
You start like this:
[like this](https://gist.github.com/tararoys/b62f1e0a29c1ffee7a93/a1ff081dd6f0f4d80217e6e23238e8ce1e3e1648)

This code will give you an error, because you have just said that a route exists.

The routes file is like a map of all of the train lines in your app.

Story: Tara goes into the customizing railcar, and draws a new rail line on the map. It doesn't exist yet, so you cant send trains there.