How do I add a post route to a rails application?
Additional questions answered in this commit:

My app says Invalid Authenticity Token when I try to submit a form. What the heck?

Comment out protect from forgery.

How do I redirect in a rails applications?

You must tell teh rails application where you are redirecting and what you are supposed to pack on the way.

A redirect is like a taxi. You give it an address and possibly a package to take with it, and it takes you to that location.